adding wtforms on messages

+++ Unit-01/09-forms/users-app/app.py
@@ -94,38 +94,48 @@ def edit(id):
 @app.route('/users/<int:user_id>/messages', methods=['GET', 'POST'])
 def messages_index(user_id):
 	#find a user
+	delete_form = DeleteForm()
 	if request.method == 'POST':
-		new_message = Message(request.form['content'], user_id)
-		db.session.add(new_message)
-		db.session.commit()
-		return redirect(url_for('messages_index', user_id=user_id))
-	
-	return render_template('messages/index.html', user=User.query.get(user_id))
+		message_form = MessageForm(request.form)
+		if message_form.validate():
+			new_message = Message(request.form['content'], user_id)
+			db.session.add(new_message)
+			db.session.commit()
+			return redirect(url_for('messages_index', user_id=user_id, form=message_form))
+		else:
+			return render_template('messages/new.html', user=User.query.get(user_id), form=message_form)
+	return render_template('messages/index.html', user=User.query.get(user_id), delete_form=delete_form)
 
 # NEW MESSAGE
 @app.route('/users/<int:user_id>/messages/new', methods=['GET','POST'])
 def messages_new(user_id):
-	message_form = MessageForm()
+	message_form = MessageForm(request.form)
 	return render_template('messages/new.html', user=User.query.get(user_id), form=message_form)
 
 #EDIT MESSAGE FORM
 @app.route('/users/<int:user_id>/messages/<int:id>/edit')
 def messages_edit(user_id, id):
 	found_message = Message.query.get(id)
-	return render_template('messages/edit.html', message=found_message)
+	message_form = MessageForm(obj=found_message)
+	return render_template('messages/edit.html', message=found_message, form=message_form)
 
 # EDIT MESSAGE
 @app.route('/users/<int:user_id>/messages/<int:id>', methods=['GET','PATCH','DELETE'])
 def messages_show(user_id, id):
 	found_message = Message.query.get(id)
 	if request.method == b'PATCH':
-		found_message.content = request.form['content']
-		db.session.add(found_message)
-		db.session.commit()
-		return redirect(url_for('messages_index', user_id=user_id))
+		message_form = MessageForm(request.form)
+		if message_form.validate():
+			found_message.content = request.form['content']
+			db.session.add(found_message)
+			db.session.commit()
+			return redirect(url_for('messages_index', user_id=user_id))
+		return render_template('messages/edit.html', message=found_message, form=message_form)
 	if request.method == b'DELETE':
-		db.session.delete(found_message)
-		db.session.commit()
+		delete_form = DeleteForm(request.form)
+		if delete_form.validate():
+			db.session.delete(found_message)
+			db.session.commit()
 		return redirect(url_for('messages_index', user_id=user_id))
 	return render_template('messages/show.html', message=found_message)
